refactor: consolidate schema files and add project/branch tables with relationships
