+++ drizzle.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig } from "drizzle-kit";
 
 export default defineConfig({
-  schema: "./src/core/db/schema/*",
+  schema: "./src/core/db/schema.ts",
   out: "./drizzle",
   dialect: "sqlite",
   dbCredentials: {
