--- src/core/db/schema/index.ts
@@ -1,1 +1,0 @@
-export * from "./intents";
